requirements.txt
Flask==3.0.3
Werkzeug==3.0.6
click==8.1.7
itsdangerous==2.2.0
Jinja2==3.1.4
pytest==8.3.3
gunicorn==23.0.0
SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.1.1
blinker==1.8.2
typing_extensions==4.12.2
MarkupSafe==3.0.2
python-dotenv==1.0.1
pandas==2.2.3
scikit-learn==1.5.2
joblib==1.4.2
numpy==2.1.3